// components/admin/adminExamTest/adminExamTestsList/adminExamTestItemStyle.scss.d.ts
export type Styles = {
  'buttonAction': string;
  'buttonActionDelete': string;
  'buttonActionEdit': string;
  'buttonsActionsContainer': string;
  'buttonShowDemoTest': string;
  'emptyExamTest': string;
  'testItemContainer': string;
  'testTitle': string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;